+++ ddsLoader.cpp
@@ -4,10 +4,10 @@ wchar_t* multiByteToWideChar(const std::string& pKey)
 {
 	const  char* pCStrKey = pKey.c_str();
 	//第一次调用返回转换后的字符串长度，用于确认为wchar_t*开辟多大的内存空间
-	int pSize = MultiByteToWideChar(CP_OEMCP, 0, pCStrKey, strlen(pCStrKey) + 1, NULL, 0);
+    int pSize = MultiByteToWideChar(CP_OEMCP, 0, pCStrKey, static_cast<int>(strlen(pCStrKey) + 1), NULL, 0);
 	wchar_t* pWCStrKey = new wchar_t[pSize];
 	//第二次调用将单字节字符串转换成双字节字符串
-	MultiByteToWideChar(CP_OEMCP, 0, pCStrKey, strlen(pCStrKey) + 1, pWCStrKey, pSize);
+    MultiByteToWideChar(CP_OEMCP, 0, pCStrKey, static_cast<int>(strlen(pCStrKey) + 1), pWCStrKey, pSize);
 	return pWCStrKey;
 }
